add new actix client
